+++ previous-versions/accounting/src/sdk/models/operations/uploadinvoiceattachment.ts
@@ -47,7 +47,7 @@ export class UploadInvoiceAttachmentResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;