+++ previous-versions/accounting/src/sdk/sdk.ts
@@ -78,9 +78,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.39.0";
-    genVersion = "2.169.0";
-    userAgent = "speakeasy-sdk/typescript 0.39.0 2.169.0 3.0.0 @codat/accounting";
+    sdkVersion = "0.40.0";
+    genVersion = "2.173.0";
+    userAgent = "speakeasy-sdk/typescript 0.40.0 2.173.0 3.0.0 @codat/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);