+++ previous-versions/accounting/RELEASES.md
@@ -699,3 +699,13 @@ Based on:
 - [typescript v0.39.0] previous-versions/accounting
 ### Releases
 - [NPM v0.39.0] https://www.npmjs.com/package/@codat/accounting/v/0.39.0 - previous-versions/accounting
+
+## 2023-10-30 01:23:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.40.0] previous-versions/accounting
+### Releases
+- [NPM v0.40.0] https://www.npmjs.com/package/@codat/accounting/v/0.40.0 - previous-versions/accounting